+++ packages/slingshot-assets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lastshotlabs/slingshot-assets
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @lastshotlabs/slingshot-entity@0.2.6
+
 ## 0.2.4
 
 ### Patch Changes