+++ packages/slingshot-chat/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lastshotlabs/slingshot-chat
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @lastshotlabs/slingshot-entity@0.2.6
+  - @lastshotlabs/slingshot-notifications@0.2.5
+  - @lastshotlabs/slingshot-push@0.2.4
+
 ## 0.2.5
 
 ### Patch Changes