+++ packages/slingshot-push/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @lastshotlabs/slingshot-push
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @lastshotlabs/slingshot-entity@0.2.6
+  - @lastshotlabs/slingshot-notifications@0.2.5
+
 ## 0.2.3
 
 ### Patch Changes