+++ packages/slingshot-webhooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lastshotlabs/slingshot-webhooks
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @lastshotlabs/slingshot-entity@0.2.6
+
 ## 0.2.3
 
 ### Patch Changes